chore: checkpoint current MCP Gateway and Ailuros Control state

+++ src/main/java/com/mcpgateway/ratelimit/SubscriptionQuotaService.java
@@ -114,7 +114,7 @@ public class SubscriptionQuotaService {
                 return QuotaCheckResult.builder()
                         .allowed(false)
                         .reason("Monthly quota exceeded")
-                        .remainingQuota(remaining)
+                        .remainingQuota(remaining.longValue())
                         .resetTime(getNextMonthResetTime())
                         .tier(tier.getName())
                         .build();
